+++ test/lnd_messages/test_open_channel_request.js
@@ -25,8 +25,10 @@ const makeArgs = overrides => {
 
 const makeExpected = overrides => {
   const args = {
+    base_fee_mtokens: undefined,
     chain_fee_tokens_per_vbyte: 1,
     cooperative_close_address: 'close_address',
+    fee_rate: undefined,
     give_tokens: 1,
     is_private: undefined,
     local_tokens: 1,