+++ test/lnd_methods/macaroon/test_handle_rpc_request_update.js
@@ -99,8 +99,10 @@ const tests = [
         id: 1,
         macaroon: undefined,
         request: {
+          base_fee_mtokens: undefined,
           chain_fee_tokens_per_vbyte: 1,
           cooperative_close_address: 'close_address',
+          fee_rate: undefined,
           give_tokens: 1,
           is_private: undefined,
           local_tokens: 1,