+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Versions
 
+## 6.4.0
+
+- `subscribeToRpcRequests`: `open_channel_request` add support for
+    `fee_rate`, `base_fee_mtokens`
+
 ## 6.3.2
 
 - `getWalletInfo`: Add support for LND 0.15.5